# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HSE Mobile Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: getQuestions
      tags: [questions]
      summary: List questions
      security:
        - bearerAuth: []
      parameters:
        - name: templateId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create question
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    put:
      operationId: updateQuestion
      tags: [questions]
      summary: Update question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /schedules:
    get:
      operationId: getSchedules
      tags: [schedules]
      summary: List inspection schedules
      security:
        - bearerAuth: []
      parameters:
        - name: supervisorId
          in: query
          schema:
            type: integer
        - name: week
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create schedule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    put:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update schedule
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete schedule
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Question:
      type: object
      properties:
        id:
          type: integer
        templateId:
          type: integer
        text:
          type: string
        answerType:
          type: string
          enum: [yes_no, text]
        isMandatory:
          type: boolean
        requiresPhoto:
          type: boolean
        categoryId:
          type: integer
        categoryName:
          type: string
        orderIndex:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - templateId
        - text
        - answerType
        - isMandatory
        - requiresPhoto
        - orderIndex
        - createdAt
    CreateQuestionRequest:
      type: object
      properties:
        templateId:
          type: integer
        text:
          type: string
        answerType:
          type: string
          enum: [yes_no, text]
        isMandatory:
          type: boolean
        requiresPhoto:
          type: boolean
        categoryId:
          type: integer
        orderIndex:
          type: integer
      required:
        - templateId
        - text
        - answerType
        - isMandatory
        - requiresPhoto
        - orderIndex
    Schedule:
      type: object
      properties:
        id:
          type: integer
        supervisorId:
          type: integer
        supervisorName:
          type: string
        templateId:
          type: integer
        templateName:
          type: string
        plantId:
          type: integer
        plantName:
          type: string
        weekStart:
          type: string
        weekEnd:
          type: string
        status:
          type: string
          enum: [pending, completed]
        inspectionId:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - supervisorId
        - supervisorName
        - templateId
        - templateName
        - plantId
        - plantName
        - weekStart
        - weekEnd
        - status
        - createdAt
    CreateScheduleRequest:
      type: object
      properties:
        supervisorId:
          type: integer
        templateId:
          type: integer
        plantId:
          type: integer
        weekStart:
          type: string
        weekEnd:
          type: string
      required:
        - supervisorId
        - templateId
        - plantId
        - weekStart
        - weekEnd
    UpdateScheduleRequest:
      type: object
      properties:
        supervisorId:
          type: integer
        templateId:
          type: integer
        plantId:
          type: integer
        weekStart:
          type: string
        weekEnd:
          type: string
        status:
          type: string
          enum: [pending, completed]
...